Use an Archon-style Aeon request
The current Aeon request approach uses a link/GET request to deliver a set of
parameters to Aeon that are used to pre-populate the request form. The updated
approach uses a form to POST the parameters - in the same way Archon was used -
so that a new Aeon request is generated without any additional user interaction.